requirements/common.txt
django-redis==4.10.0
redis==2.10.6
futures
gunicorn==19.10.0
Django==1.11.6
lxml
pymysql
mysqlclient==2.0.3
hurry.filesize==0.9
pysolr==3.9.0
mock==2.0.0
django-autocomplete-light==2.3.3
sorl-thumbnail==12.3
Pillow
psycopg2==2.7.3.2
xlwt==1.3.0
Unidecode==0.4.21
unicodecsv==0.14.1
openpyxl==2.4.9
django-allauth==0.33.0
django-simple-captcha==0.5.5
django-haystack
haversine
libsass==0.13.3
django-appconf==1.0.3
django-compressor==2.2
django-sass-processor
django-storages==1.7.1
six==1.10.0
Levenshtein_search
ijson==2.6.1
requests==2.24.0
mysql-connector-python==8.0.26
protobuf==3.19.1
pandas